deletes controller class
