+++ src/test/java/de/rabea/server/HttpServerTest.java
@@ -73,4 +73,13 @@ public class HttpServerTest {
 
         assertEquals("HTTP/1.1 200 OK\n\n" + "GET /logs HTTP/1.1\n", networkStub.returnedResponse);
     }
+
+    @Test
+    public void returnsUrlParamsInTheResponse() {
+        NetworkStub networkStub = new NetworkStub ("GET /parameters?code=hello HTTP/1.1\n");
+        HttpServer httpServer = new HttpServer(networkStub, new ContentStorage(), log);
+        httpServer.start(directory);
+
+        assertEquals("HTTP/1.1 200 OK\n\n" + "code = hello\n", networkStub.returnedResponse);
+    }
 }
